+++ lib/roda/data_grid/data_grid_helpers.rb
@@ -46,6 +46,7 @@ class Roda
           query_param.build_list.list_values
         else
           query_param.build_list do |sql|
+            raise "SQL for #{param list} is not set" if sql.nil?
             raise "SQL for #{param list} is not a SELECT" if sql.match?(/insert |update |delete /i)
 
             connection[sql].map(&:values)